lir: a constructor builds the operand-carrying instructions
`BinOp`, `Compare` and `UnaryOp` were spelled as struct literals at 84
sites, so a change to what they carry is a change at every one of them.
`LirInstr::binop`/`compare`/`unary` take the same operands and give the
representation one place to change.

Auditing the files that touched brought two of them under the reading
budget: the constructor tests leave `lower/pattern.rs` for `pattern/ctor.rs`,
and the constant walk leaves `types/instr.rs` for `instr/consts.rs`.
`types/instr.rs` names one enum the language cannot split, so
`src/lir/AGENTS.md` records it under the dispatch-table allowance the root
index already grants.
